requirements.txt
aiohttp==1.3.3
Pillow==4.0.0
python-magic==0.4.12
requests==2.5.1
